Updated code for create customer payment profile
Fixes the response issue in create custome payment profile.
The response was not having customer profile id and newly created customer payment profile id. The request that was being sent was for get customer profile which had a different response format. The request is changed and now correct response is obtained.

Signed-off-by: Srijan Misra <srmisra@visa.com>